Derive 'node_number' field from MAC addresses
The node_number was previously derived from a node's management IP.
In a routed rack configuration, this could cause the node_number to
wrap around and overlap inside a single cluster.

This commit derives the node_number from the MAC address on the
management interface.  The MAC is converted to an integer.  The node
number is this integer modulo Integer.MAX_VALUE.

This gives us a ~31 bit positive integer that is guaranteed to be
unique inside any single-vendor customer.  It is very likely to be
unique across any two NICs, but not certainly so, since only the
last 7 bits of a 24 bit OUI are present.

Zookeeper and keepalived use (node_number % 8) to derive an 8 bit
value from the 31 bit node number.  In a sufficiently large cluster
this may overlap.  A helper method added to libraries/node_numbers.rb
does a node search to raise an exception in the event of an overlap.