Suppress Hashie warnings once and for all
Hashie, a transitive dependency buried inside Chef, added a feature that warns the user each and every time a stdlib method might be overridden by a generated accessor method.  This fills our C-A-R logs with useless/trivial garbage:

```
W, [2017-09-07T20:16:20.484655 #18995]  WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#size defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.
W, [2017-09-07T20:16:20.486006 #18995]  WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#size defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.
W, [2017-09-07T20:16:20.487087 #18995]  WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#size defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.
W, [2017-09-07T20:16:20.488035 #18995]  WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#size defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.
W, [2017-09-07T20:16:20.489211 #18995]  WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#size defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.
```

This PR loads Hashie early, then suppresses the warnings.