Track all changes in cookbooks/
We have several problems in upgrade when there are local changes in
`cookbooks/`.  If there is a `Gemfile` in a cookbook, it mutates the
overall `Gemfile` that bundler detects.  This then fails the bootstrap's
chef-client.  This error is hard to detect and figure out.  Having this
should give better clues.